Keep the coordinate reader in the coordinate module
_coordinate_values sat in _timeseries.py while _coords.py held everything
else about coordinates. Moving it also lets _coords.py read coordinates
without importing _timeseries.py, which already imports _coords.py.
